api/requirements.txt
click==7.1.2
fastapi==0.61.1
h11==0.10.0
mysql-connector-python==8.0.21
protobuf==3.13.0
pydantic==1.6.1
six==1.15.0
starlette==0.13.6
uvicorn==0.12.1
